moving markdown components to components
