+++ src/app/(api)/api/[...slug]/page.tsx
@@ -1,8 +1,7 @@
-import { ApiRequest, ApiResponses, TimeAgo } from '@/components';
+import { ApiRequest, ApiResponses, TimeAgo, Heading, Paragraph } from '@/components';
 import { loadApiContentInfo } from '@/lib/markdown/util';
 import { parseSchema, toOperations } from '@/lib/swagger/parse';
 import { toRouteSegments, toSlug } from '@/lib/util';
-import { Heading, Paragraph } from '@/markdown';
 import { notFound } from 'next/navigation';
 import type { Metadata } from 'next';
 import { withMdxMetadata, withDefaultMetadata, getLastUpdated } from '@/lib/metadata/util';
