+++ src/app/(documentation)/design-system/images/horizontal.png
