+++ src/app/(documentation)/design-system/images/vertical.png
