+++ src/app/(documentation)/design-system/page.module.css
@@ -1,0 +1,21 @@
+.iconGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
+  gap: var(--space-m);
+  margin: var(--space-m) 0;
+}
+
+.iconWrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: var(--space-m);
+  border: 1px solid var(--color-border);
+  border-radius: var(--border-radius-m);
+}
+
+.iconName {
+  margin-top: var(--space-xs);
+  font-size: var(--text-body-xs);
+  text-align: center;
+}
