--- src/app/(internal)/design-system/page.module.css
@@ -1,59 +1,0 @@
-.section {
-  padding: var(--space-2xl) 0;
-}
-
-.container {
-  max-width: var(--container-width-xl);
-  padding: 0 var(--space-m);
-  margin: 0 auto;
-}
-
-.sectionHeading {
-  font-family: var(--font-family-heading);
-  margin-top: var(--space-m);
-  margin-bottom: var(--space-m);
-}
-
-.sectionDescription {
-  margin-bottom: var(--space-m);
-}
-
-.subHeading {
-  font-family: var(--font-family-heading);
-  margin-top: var(--space-m);
-  margin-bottom: var(--space-m);
-}
-
-.componentContainer {
-  margin-bottom: var(--space-3xl);
-}
-
-.flexContainer {
-  margin-bottom: var(--space-3xl);
-}
-
-.videoContainer {
-  max-width: var(--container-width-s);
-}
-
-.iconGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
-  gap: var(--space-m);
-  margin: var(--space-m) 0;
-}
-
-.iconWrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: var(--space-m);
-  border: 1px solid var(--color-border);
-  border-radius: var(--border-radius-m);
-}
-
-.iconName {
-  margin-top: var(--space-xs);
-  font-size: var(--text-body-xs);
-  text-align: center;
-}
