--- src/app/(internal)/design-system/page.tsx
@@ -1,330 +1,0 @@
-import { BlockImage, Button, Card, Flex, HomepageHero, Navbar, Note, Video } from '@/components';
-import horizontalImage from '@/content/test/images/horizontal.png';
-import verticalImage from '@/content/test/images/vertical.png';
-import { Icon, IconName } from '@/icons';
-import { cx } from 'class-variance-authority';
-import { iconRegistry } from '../../../icons/icon-registry';
-
-import styles from './page.module.css';
-
-export default function DesignSystemPage() {
-  return (
-    <div className={styles.root}>
-      <section id="buttons" className={cx(styles.section, styles.container)}>
-        <h2 className={styles.sectionHeading}>Buttons</h2>
-
-        <h3>Primary Buttons</h3>
-        <Flex gap="m" align="center" className={styles.flexContainer}>
-          <Button href="/" size="small">
-            Small Button
-          </Button>
-          <Button href="/">Medium Button</Button>
-          <Button href="/" size="large">
-            Large Button
-          </Button>
-          <Button href="/" size="xlarge">
-            XLarge Button
-          </Button>
-        </Flex>
-
-        <h3 className={styles.subHeading}>Secondary Buttons</h3>
-        <Flex gap="m" align="center" className={styles.flexContainer}>
-          <Button href="/" variant="secondary" size="small">
-            Small Button
-          </Button>
-          <Button href="/" variant="secondary">
-            Medium Button
-          </Button>
-          <Button href="/" variant="secondary" size="large">
-            Large Button
-          </Button>
-          <Button href="/" variant="secondary" size="xlarge">
-            XLarge Button
-          </Button>
-        </Flex>
-
-        <h3 className={styles.subHeading}>Link Buttons</h3>
-        <Flex gap="m" align="center" className={styles.flexContainer}>
-          <Button href="/" variant="link" size="small">
-            Small Link
-          </Button>
-          <Button href="/" variant="link">
-            Medium Link
-          </Button>
-          <Button href="/" variant="link" size="large">
-            Large Link
-          </Button>
-          <Button href="/" variant="link" size="xlarge">
-            XLarge Link
-          </Button>
-        </Flex>
-
-        <h3 className={styles.subHeading}>Color Schemes</h3>
-        <Flex gap="m" align="center" className={styles.flexContainer}>
-          <Button href="/" colorScheme="accent">
-            Accent Button
-          </Button>
-          <Button href="/" colorScheme="text">
-            Text Button
-          </Button>
-          <Button href="/" colorScheme="dark">
-            Dark Button
-          </Button>
-          <Button href="/" colorScheme="light">
-            Light Button
-          </Button>
-        </Flex>
-
-        <h3 className={styles.subHeading}>With Arrow</h3>
-        <Flex gap="m" align="center" className={styles.flexContainer}>
-          <Button href="/" withArrow>
-            Primary with Arrow
-          </Button>
-          <Button href="/" variant="secondary" withArrow>
-            Secondary with Arrow
-          </Button>
-          <Button href="/" variant="link" withArrow>
-            Link with Arrow
-          </Button>
-        </Flex>
-
-        <h3 className={styles.subHeading}>Widths</h3>
-        <Flex gap="m" align="center" className={styles.flexContainer}>
-          <Button href="/" width="small">
-            Small
-          </Button>
-          <Button href="/" width="medium">
-            Medium
-          </Button>
-          <Button href="/" width="large">
-            Large
-          </Button>
-          <Button href="/" width="full">
-            Full Width Button
-          </Button>
-        </Flex>
-      </section>
-
-      <section id="icons" className={cx(styles.section, styles.container)}>
-        <h2 className={styles.sectionHeading}>Icons</h2>
-
-        <div className={styles.iconGrid}>
-          {Object.keys(iconRegistry).map((iconName) => (
-            <div key={iconName} className={styles.iconWrapper}>
-              <Icon name={iconName as IconName} as="svg" title={iconName} width={24} height={24} />
-              <span className={styles.iconName}>{iconName}</span>
-            </div>
-          ))}
-        </div>
-
-        <Note>
-          Some icons appear twice because they can be animated and need to be rendered separately for each
-          direction.
-        </Note>
-
-        <Note variant="tip">
-          Some icons appear twice because they can be animated and need to be rendered separately for each
-          direction.
-        </Note>
-
-        <Note variant="important">
-          Some icons appear twice because they can be animated and need to be rendered separately for each
-          direction.
-        </Note>
-
-        <Note variant="warning">
-          Some icons appear twice because they can be animated and need to be rendered separately for each
-          direction.
-        </Note>
-
-        <Note variant="caution">
-          Some icons appear twice because they can be animated and need to be rendered separately for each
-          direction.
-        </Note>
-      </section>
-
-      <section id="video" className={cx(styles.section, styles.container)}>
-        <h2 className={styles.sectionHeading}>Video</h2>
-        <p className={styles.sectionDescription}>
-          A responsive Wistia video player that adapts to its container width.
-        </p>
-
-        <h3>Default</h3>
-        <div className={styles.videoContainer}>
-          <Video wistiaId="r5d3j2nz4m" />
-        </div>
-
-        <h3>Poster Image</h3>
-        <div className={styles.videoContainer}>
-          <Video
-            wistiaId="oifkgmxnkb"
-            posterImage="https://i.vimeocdn.com/video/1666844499-1666844499_1280x720.jpg"
-          />
-        </div>
-
-        <h3>Resumable</h3>
-        <div className={styles.videoContainer}>
-          <Video wistiaId="0x8do0ucl8" resumable />
-        </div>
-      </section>
-
-      <section id="navbar" className={styles.section}>
-        <h2 className={cx(styles.sectionHeading, styles.container)}>Navbar</h2>
-
-        <Navbar />
-      </section>
-
-      <section id="homepageHero" className={styles.section}>
-        <h2 className={cx(styles.sectionHeading, styles.container)}>Homepage Hero</h2>
-
-        <HomepageHero
-          title="Welcome to Cloudsmith's documentation site."
-          description="Cloudsmith is a cloud-native, hosted, package management service with deep focus on providing the best universal support for all native package and container technologies."
-          buttons={[
-            { label: 'Documentation', href: '/' },
-            { label: 'Guides', href: '/' },
-            { label: 'API Reference', href: '/' },
-          ]}
-        />
-      </section>
-
-      <section id="cards" className={cx(styles.section, styles.container)}>
-        <h2 className={styles.sectionHeading}>Card & Flex Components Test</h2>
-
-        <Flex gap="l" align="stretch" className={styles.componentContainer}>
-          <Card
-            title="Getting Started Guide"
-            description="This is a Card with a description and default size (m)"
-            href="/"
-            linkText="Go to guide"
-            icon="utility/documentation"
-          />
-
-          <Card
-            title="Setting Your Security Policies Guide"
-            description="Dorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio mattis."
-            href="/"
-            linkText="Go to guide"
-            icon="utility/guide"
-          />
-
-          <Card
-            title="Integrating your preferred package format with Cloudsmith"
-            description="Dorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio mattis."
-            href="/"
-            linkText="Go to documentation"
-            icon="utility/guide"
-          />
-
-          <Card
-            title="Integrating your preferred package format with Cloudsmith"
-            description="Dorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio mattis."
-            href="/"
-            linkText="Go to documentation"
-            icon="utility/guide"
-          />
-        </Flex>
-
-        <h2 className={styles.sectionHeading}>Cards: Size variant (large)</h2>
-
-        <Flex gap="l" align="stretch" className={styles.flexContainer}>
-          <Card
-            title="Dependencies"
-            href="/"
-            linkText="For Developers"
-            icon="utility/documentation"
-            size="l"
-          />
-
-          <Card title="Deployment" href="/" linkText="For Operations" icon="utility/guide" size="l" />
-
-          <Card title="Distribution" href="/" linkText="For Vendors" icon="utility/api" size="l" />
-        </Flex>
-
-        <h2 className={styles.sectionHeading}>Cards: Width variants</h2>
-
-        <h3 className={styles.subHeading}>Thirds</h3>
-
-        <Flex gap="l" align="stretch" className={styles.flexContainer}>
-          <Card
-            title="Dependencies"
-            href="/"
-            linkText="For Developers"
-            icon="utility/documentation"
-            size="l"
-            width="third"
-          />
-
-          <Card
-            title="Deployment"
-            href="/"
-            linkText="For Operations"
-            icon="utility/guide"
-            size="l"
-            width="third"
-          />
-
-          <Card
-            title="Distribution"
-            href="/"
-            linkText="For Vendors"
-            icon="utility/api"
-            size="l"
-            width="third"
-          />
-        </Flex>
-
-        <h3 className={styles.subHeading}>Halves</h3>
-
-        <Flex gap="l" align="stretch" className={styles.flexContainer}>
-          <Card
-            title="Dependencies"
-            href="/"
-            linkText="For Developers"
-            icon="utility/documentation"
-            size="l"
-            width="half"
-          />
-
-          <Card
-            title="Deployment"
-            href="/"
-            linkText="For Operations"
-            icon="utility/guide"
-            size="l"
-            width="half"
-          />
-        </Flex>
-
-        <h3 className={styles.subHeading}>Full</h3>
-
-        <Flex gap="l" align="stretch" className={styles.flexContainer}>
-          <Card
-            title="This is a full width card"
-            href="/"
-            linkText="Go to documentation"
-            icon="utility/documentation"
-            size="l"
-            width="full"
-          />
-        </Flex>
-      </section>
-
-      <section id="images" className={cx(styles.section, styles.container)}>
-        <h2 className={cx(styles.sectionHeading, styles.container)}>Images</h2>
-
-        <BlockImage src={horizontalImage} alt="Image alternative text">
-          This is a caption that I would like to show for the image
-        </BlockImage>
-
-        <BlockImage align="left" src={verticalImage} alt="Image alternative text">
-          This is a caption that I would like to show for the image
-        </BlockImage>
-
-        <BlockImage align="right" src={verticalImage} alt="Image alternative text">
-          This is a caption that I would like to show for the image
-        </BlockImage>
-      </section>
-    </div>
-  );
-}
