+++ src/components/CodeBlock/CodeBlock.module.css
@@ -59,7 +59,7 @@
   position: relative;
   user-select: none;
   color: var(--color-text-on-color);
-  border-radius: 12px 12px 0 0;
+  border-radius: var(--border-radius-2xl) var(--border-radius-2xl) 0 0;
   font-family: var(--font-family-mono);
   text-transform: uppercase;
   font-size: 12px;
@@ -70,5 +70,6 @@
 }
 
 .hideHeader {
-  background-color: red;
+  border-top-left-radius: var(--border-radius-2xl);
+  border-top-right-radius: var(--border-radius-2xl);
 }
