+++ src/components/HorizontalRule/index.ts
