+++ src/components/index.ts
@@ -16,3 +16,11 @@ export * from './ApiRequest';
 export * from './ApiGrid';
 export * from './ApiMedia';
 export * from './TimeAgo';
+export * from './Code';
+export * from './CodeBlock';
+export * from './Heading';
+export * from './HorizontalRule';
+export * from './Link';
+export * from './List';
+export * from './Paragraph';
+export * from './Table';
