--- src/content/test/another-test.mdx
@@ -1,3 +1,0 @@
-# Another test
-
-This is another test.
