--- src/markdown/index.ts
@@ -1,7 +1,0 @@
-export * from './Code';
-export * from './CodeBlock';
-export * from './Heading';
-export * from './HorizontalRule';
-export * from './Link';
-export * from './List';
-export * from './Paragraph';
